requirements.txt
environs==9.3.1
marshmallow
Pillow
python-dotenv==0.15.0
